+++ smart-contracts/contracts/dex-router/src/tests/mod.rs
@@ -1,3 +1,1 @@
-mod helpers;
-mod initialize;
-mod multipool_paths;
+mod swap;
